fix: cast as decimal instead of as int

+++ sql/semester_averages.sql
@@ -32,7 +32,7 @@ COPY (
             SELECT  month,
                     COUNT(*) AS rides_per_month,
                     SUM(ride_duration_minutes) AS time_ridden,
-                    SUM(CAST(cleaned_fare_euros AS INT)) AS saved_money_euros
+                    SUM(CAST(cleaned_fare_euros AS DECIMAL(10,2))) AS saved_money_euros
             FROM cleaned_fare_euros
             GROUP BY month
         ),
